fix collation

+++ experiments/exp70/collate_samples.py
@@ -11,6 +11,7 @@ def main(
     tag: str = typer.Option(),
     mode: str = typer.Option(),
     start_string: str = typer.Option(),
+    overwrite: bool = False,
 ):
     assert mode in ("plus", "minus", "")
     mode_txt = f"_{mode}" if mode else ""
@@ -18,7 +19,10 @@ def main(
     newpath = dirpath / f"{start_string}{mode_txt}.npz"
 
     if newpath.exists():
-        os.remove(newpath)
+        if overwrite:
+            os.remove(newpath)
+        else:
+            raise FileExistsError("File already exists that is being collated.")
 
     # first collect all relevant files
     fps = []
